+++ integration/spark/app/src/main/java/io/openlineage/spark/agent/lifecycle/Spark32DatasetBuilderFactory.java
@@ -30,7 +30,9 @@ import io.openlineage.spark3.agent.lifecycle.plan.SubqueryAliasInputDatasetBuild
 import io.openlineage.spark3.agent.lifecycle.plan.TableContentChangeDatasetBuilder;
 import io.openlineage.spark32.agent.lifecycle.plan.AlterTableCommandDatasetBuilder;
 import io.openlineage.spark32.agent.lifecycle.plan.column.MergeIntoDelta11ColumnLineageVisitor;
+import io.openlineage.spark32.agent.lifecycle.plan.column.MergeIntoIceberg013ColumnLineageVisitor;
 import io.openlineage.spark34.agent.lifecycle.plan.column.MergeIntoDelta24ColumnLineageVisitor;
+import io.openlineage.spark34.agent.lifecycle.plan.column.MergeIntoIceberg13ColumnLineageVisitor;
 import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.List;
@@ -82,18 +84,6 @@ public class Spark32DatasetBuilderFactory implements DatasetBuilderFactory {
     return builder.build();
   }
 
-  private boolean hasAlterTableClass() {
-    try {
-      Spark32DatasetBuilderFactory.class
-          .getClassLoader()
-          .loadClass("org.apache.spark.sql.catalyst.plans.logical.AlterTable");
-      return true;
-    } catch (Exception e) {
-      // swallow- we don't care
-    }
-    return false;
-  }
-
   private AbstractQueryPlanOutputDatasetBuilder<LogicalPlan> getCreateReplaceDatasetBuilder(
       OpenLineageContext context) {
     // On Databricks platform Spark 3.2 needs to have the version of `CreateReplaceDatasetBuilder`
@@ -125,6 +115,14 @@ public class Spark32DatasetBuilderFactory implements DatasetBuilderFactory {
 
     if (MergeIntoDelta11ColumnLineageVisitor.hasClasses()) {
       builder.add(new MergeIntoDelta11ColumnLineageVisitor(context));
+    }
+
+    if (MergeIntoIceberg13ColumnLineageVisitor.hasClasses()) {
+      builder.add(new MergeIntoIceberg13ColumnLineageVisitor(context));
+    }
+
+    if (MergeIntoIceberg013ColumnLineageVisitor.hasClasses()) {
+      builder.add(new MergeIntoIceberg013ColumnLineageVisitor(context));
     }
 
     return builder.build();