+++ integration/spark/shared/src/main/java/io/openlineage/spark/agent/util/BigQueryUtils.java
@@ -5,7 +5,14 @@
 
 package io.openlineage.spark.agent.util;
 
+import static io.openlineage.spark.agent.util.ReflectionUtils.tryExecuteMethod;
+
+import com.google.cloud.spark.bigquery.BigQueryRelation;
 import io.openlineage.spark.agent.lifecycle.plan.BigQueryNodeOutputVisitor;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Stream;
 
 public class BigQueryUtils {
 
@@ -19,5 +26,30 @@ public class BigQueryUtils {
       // swallow- we don't care
     }
     return false;
+  }
+
+  private static Optional<Object> extractDatasetIdentifierFromTableId(Object tableId) {
+    return Stream.of(
+            ReflectionUtils.tryExecuteStaticMethodForClassName(
+                "com.google.cloud.bigquery.connector.common.BigQueryUtil",
+                "friendlyTableName",
+                tableId),
+            ReflectionUtils.tryExecuteStaticMethodForClassName(
+                "com.google.cloud.spark.bigquery.repackaged.com.google.cloud.bigquery.connector.common.BigQueryUtil",
+                "friendlyTableName",
+                tableId))
+        .filter(Optional::isPresent)
+        .map(Optional::get)
+        .findFirst();
+  }
+
+  public static List<DatasetIdentifier> extractDatasetIdentifier(
+      BigQueryRelation bigQueryRelation) {
+
+    return tryExecuteMethod(bigQueryRelation, "getTableId")
+        .flatMap(BigQueryUtils::extractDatasetIdentifierFromTableId)
+        .map(x -> new DatasetIdentifier((String) x, "namespace"))
+        .map(Collections::singletonList)
+        .orElseGet(Collections::emptyList);
   }
 }